Removing old MERF and replacing with new API.
